chore: refactor banners pt.1
