+++ src/components/banners/HomeBanners.jsx
@@ -11,7 +11,7 @@ import { connectQueryRules } from 'react-instantsearch-dom';
 import { Link } from 'react-router-dom';
 
 // This component renders a different banner based on the props passed to it.
-//The props are passed through the Dashboard in rules section.
+// The props are passed through the Dashboard in rules section.
 const HomeBanners = ({ items }) => {
   return items.map(
     (
@@ -29,35 +29,41 @@ const HomeBanners = ({ items }) => {
       },
       index
     ) => {
+      if (type === 'HomeBannerOne') {
+        return (
+          <Fragment key={index}>
+            <BannerOne
+              index={index}
+              title={title}
+              subtitle={subtitle}
+              imgUrl2={imgUrl2}
+              imgUrl3={imgUrl3}
+              LinkButton1={LinkButton1}
+              button1={button1}
+              LinkButton2={LinkButton2}
+              button2={button2}
+            />
+        </Fragment>
+        );
+      }
       if (type === 'HomeBannerTwo') {
         return (
-          <div
-            className="home-banner2-container"
-            style={{
-              backgroundImage: `linear-gradient(rgba(0, 0, 0, 0.62), rgba(0, 0, 0, 0.45)), url(${imgUrl1})`,
-            }}
-            key={index}
-          >
-            <div className="home-banner2-container__infos">
-              <h1>{title}</h1>
-              <h2>{subtitle}</h2>
-            </div>
-            <div className="home-banner2-container__buttons">
-              <div className="home-banner2-container__buttons__circles">
-                <div className="home-banner2-container__buttons__circles__circles2">
-                  <div className="home-banner2-container__buttons__circles__circles2__circles3">
-                    <Link to={LinkButton1}>{button1}</Link>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
+          <Fragment>
+            <BannerTwo
+              index={index}
+              title={title}
+              subtitle={subtitle}
+              LinkButton1={LinkButton1}
+              button1={button1}
+            />
+        </Fragment>
         );
       }
       if (type === 'HomeBannerThree') {
         return (
           <Fragment key={index}>
             <BannerThree
+              index={index}
               imgUrl1={imgUrl1}
               title={title}
               subtitle={subtitle}
@@ -67,35 +73,58 @@ const HomeBanners = ({ items }) => {
           </Fragment>
         );
       }
-      if (type === 'HomeBannerOne') {
-        return (
-          <div
-            className="home-banner-container"
-            style={{
-              backgroundImage: `linear-gradient(rgba(0, 0, 0, 0.62), rgba(0, 0, 0, 0.45)), url(${imgUrl1})`,
-            }}
-            key={index}
-          >
-            <div className="home-banner-container__infos">
-              <h1>{title}</h1>
-              <h2>{subtitle}</h2>
-            </div>
-            <div className="home-banner-container__image-one">
-              <img src={imgUrl2} alt="" />
-            </div>
-            <div className="home-banner-container__image-two">
-              <img src={imgUrl3} alt="" />
-              <div className="home-banner-container__buttons">
-                <Link to={LinkButton1}>{button1}</Link>
-                <Link to={LinkButton2}>{button2}</Link>
-              </div>
-            </div>
-          </div>
-        );
-      }
     }
   );
 };
+
+const BannerOne = ({title, subtitle, imgUrl2, imgUrl3, LinkButton1, button1, LinkButton2, button2}) => (
+  <div
+    className="home-banner-container"
+    style={{
+      backgroundImage: `linear-gradient(rgba(0, 0, 0, 0.62), rgba(0, 0, 0, 0.45)), url(${imgUrl1})`,
+    }}
+    key={index}
+  >
+    <div className="home-banner-container__infos">
+      <h1>{title}</h1>
+      <h2>{subtitle}</h2>
+    </div>
+    <div className="home-banner-container__image-one">
+      <img src={imgUrl2} alt="" />
+    </div>
+    <div className="home-banner-container__image-two">
+      <img src={imgUrl3} alt="" />
+      <div className="home-banner-container__buttons">
+        <Link to={LinkButton1}>{button1}</Link>
+        <Link to={LinkButton2}>{button2}</Link>
+      </div>
+    </div>
+  </div>
+)
+
+const BannerTwo = ({title, subtitle, LinkButton1, button1}) => (
+  <div
+    className="home-banner2-container"
+    style={{
+      backgroundImage: `linear-gradient(rgba(0, 0, 0, 0.62), rgba(0, 0, 0, 0.45)), url(${imgUrl1})`,
+    }}
+    key={index}
+  >
+    <div className="home-banner2-container__infos">
+      <h1>{title}</h1>
+      <h2>{subtitle}</h2>
+    </div>
+    <div className="home-banner2-container__buttons">
+      <div className="home-banner2-container__buttons__circles">
+        <div className="home-banner2-container__buttons__circles__circles2">
+          <div className="home-banner2-container__buttons__circles__circles2__circles3">
+            <Link to={LinkButton1}>{button1}</Link>
+          </div>
+        </div>
+      </div>
+    </div>
+</div>
+)
 
 const BannerThree = ({ imgUrl1, title, subtitle, LinkButton1, button1 }) => {
   return (
